requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
mysqlclient==2.2.5
numpy==2.1.2
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.2
Werkzeug==3.0.4
gunicorn==23.0.0
